+++ packages/amplify-provider-awscloudformation/src/resourceParams.js
@@ -43,7 +43,7 @@ function loadResourceParameters(context, category, resource) {
   const envSpecificParams = context.amplify.loadEnvResourceParameters(context, category, resource);
   let resourceParameters = { ...parameters, ...envSpecificParams };
   if (category === 'auth' && parameters && parameters.hostedUI && !resourceParameters[hostedUIProviderCredsField]) {
-    resourceParameters = _.set(resourceParameters, hostedUIProviderCredsField, '[]');
+    resourceParameters = _.setWith(resourceParameters, hostedUIProviderCredsField, '[]');
   }
   return resourceParameters;
 }