+++ packages/amplify-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@4.3.10...amplify-app@4.3.11) (2023-02-27)
+
+**Note:** Version bump only for package amplify-app
+
+
+
+
+
 ## [4.3.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@4.3.9...amplify-app@4.3.10) (2023-02-16)
 
 **Note:** Version bump only for package amplify-app