+++ packages/amplify-function-plugin-interface/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-function-plugin-interface@1.9.6...amplify-function-plugin-interface@1.9.7) (2023-02-27)
+
+**Note:** Version bump only for package amplify-function-plugin-interface
+
+
+
+
+
 ## [1.9.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-function-plugin-interface@1.9.5...amplify-function-plugin-interface@1.9.6) (2023-01-03)
 
 **Note:** Version bump only for package amplify-function-plugin-interface